Fix id of tags and Finalize the page
Added short description of the page under the description heading
and remove the old description heading. Modified the directions
heading to be the steps heading and added ordered list of steps to take
when using the recipe. Added two links to other recipes on the internet
at the bottom of the page.